HHH-18139 isolate generator creation in GeneratorBinder
